// src/button/interface.ts
export interface ButtonProps {
  prefixCls?: string;
  style?: any;
  className?: string;
  href?: string;
  type?: string;
  size?: string;
  loading?: Boolean;
  block?: Boolean;
  disabled?: Boolean;
  hollow?: Boolean;
  dashed?: Boolean;
  circle?: Boolean;
  plain?: Boolean;
  htmlType?: string;
  types?: any;
  onClick?: () => void;
}